+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- Default to mapped table name instead of model name for identifying prisma models.  
+
 ## 0.0.4 (2023-08-23)
 
 - Add support for reading prisma schema configuration from `package.json`.